refactor: move StaleBlock to the shared components (#235)
