+++ src/components/common/StaleBlock.tsx
@@ -21,6 +21,11 @@ import { cn } from '@/lib/utils/cn'
   ▸ **못 누르게 한다** — 옛 행을 눌러 다른 회차의 상세로 가는 사고를 막는다.
   ▸ **말한다** — 「불러오는 중」. 흐림은 눈으로만 보이는 신호라 `aria-busy`도 같이 단다.
 
+  ─── 왜 공용인가 ───────────────────────────────────────────────
+  오퍼레이터가 먼저 만들었고 **매니저가 두 번째로 필요해져 올렸다**(D14 — 도메인이
+  갈리면 두 번째에 올린다). 레이어 린트가 `features` 간 교차 import를 막으므로 여기
+  말고는 둘 다 쓸 자리가 없다.
+
   ─── 스켈레톤과 갈리는 지점 ─────────────────────────────────────
   **처음 오는 것은 스켈레톤**(보여줄 옛 값이 없다), **바뀌는 것은 이것**(옛 값이 있다).
   둘을 바꿔 쓰면 조건을 바꿀 때마다 표가 통째로 사라졌다 다시 생긴다.
