+++ src/features/operator/projects/list/ProjectListScreen.tsx
@@ -16,7 +16,7 @@ import ErrorState from '@/components/common/ErrorState'
 import TableSkeleton from '@/components/common/TableSkeleton'
 import { cn } from '@/lib/utils/cn'
 import { useDebounced } from '@/lib/useDebounced'
-import StaleBlock from '../../_shared/StaleBlock'
+import StaleBlock from '@/components/common/StaleBlock'
 import { getToday, useCohortScope, useLinkableCurricula, useProjectList } from '../queries'
 import { CONCEPT_COUNT, dueLabel } from '../rules'
 import { useCohortId } from '@/stores/cohortScope'
